Added ability to distinguish between environments
This pull request makes it possible to distinguish between `test` and `demo`
environments when deployed to Amazon Elastic Beanstalk.

Because of the (current) restrictions of Beanstalk, we leverage one of its
unnamed environment variables (`PARAM1`) to denote if there is something
special about this instance.

This variable can be set within the Beanstalk web console in the **Container**
tab.

It is expected that `PARAM1` be set to either `TEST` or `DEMO`. If it is *not*
set, then it is assumed that we are merely operating on a production instance,
as expected.

In the future, should Beanstalk become more flexible, we can use a more
reasonable variable name such as `DEFAULT_ENV` or something else suitable.

For now, the new settings files operate as follows:
- `test.py` inherits from the `dev.py` file
- `demo.py` inherits from the `production.py` file

No changes have been made to handle different requirements.