Prevent double pageview tracking
We only track pageviews in the following locations:
1. `pages.tracker.js:setup`
2. `pages.tracker.js:trackEvent` > only on scroll
3. `pages.tracker.js:defaultEventMap` > `click .tile`
4. `pages.widget.gallery.js` > on image click

If the problem were case 2, we would see `/livedin/#page{scrolls}` in the Google Analytics Debugger; this is not observed
If the problem were case 3, we would see `/livedin/#{tile_id}` in the Google Analytics Debugger; this is not observed
If the problem were case 4, we would see `/livedin/#{tile_id}&photos={id}` in the Google Analytics Debugger; this is not observed

Another possibility is that Google Analytics is being loaded multiple times leading to it being tracked multiple times. Looking through the theme and the code, `analytics.js` is found in exactly one place, `pages.tracker.js`, and the only reference to `ga.js` the previous analytics, is only found in `SpecRunner.html` (unrelated).

In which case, it is possible that out tracker is being initialized twice. By observation, this is not the case.

All of this leads to case 1 being the cuplrit. Further to this suspicion, according to [the docs](https://developers.google.com/analytics/devguides/collection/analyticsjs/pages):

> The default JavaScript snippet for analytics.js contains a page tracking call, so on most sites, page tracking is already happening through the default snippet.

If it is not case 1 (i.e. the fix in this commit) then, frankly, I have no idea.