Theory: Seperate page views contribute to bounce
For reference, from what I've read:

    Bounce Rate = % of single-page visits with only one event

Because of this, its possible that whenever a person enters / leaves a pop-up, there is only one event tracked so we see a high bounce rate.

That being said, if I am wrong, this code wasn't really contributing much anyway; it was asked to be added on a whim and laregely hasn't been used.