Added a monkey-patch for `extract_first`
Throughout the Scrapy code, we often only `extract` one element, but we always get a `SelectorList`. This is great sometimes, but a lot of the time, we only want one element.

The proper solution would be to submit a PR against the Scrapy project, and use our own fork of it in the meantime, but the approach I've implemented should ruffle less feathers.